test : 내 쉐어 조회 API 실패케이스 통합 테스트 구현
