+++ src/test/java/louie/hanse/shareplate/integration/share/ShareIntegrationTest.java
@@ -74,96 +74,6 @@ class ShareIntegrationTest extends InitIntegrationTest {
     }
 
     @Test
-    void 내가_신청한_쉐어를_조회한다() {
-        String accessToken = jwtProvider.createAccessToken(2355841047L);
-
-        given(documentationSpec)
-            .filter(document("share-search-mine-entry-get"))
-            .accept(APPLICATION_JSON_VALUE)
-            .header(AUTHORIZATION, accessToken)
-            .param("mineType", "entry")
-            .param("shareType", "delivery")
-            .param("isExpired", "false")
-
-            .when()
-            .get("/shares/mine")
-
-            .then()
-            .statusCode(HttpStatus.OK.value())
-            .body("", hasSize(1))
-            .body("[0].id", equalTo(1))
-            .body("[0].thumbnailUrl", containsString("https://"))
-            .body("[0].title", equalTo("강남역에서 떡볶이 먹을 사람 모집합니다."))
-            .body("[0].location", equalTo("강남역"))
-            .body("[0].price", equalTo(10000))
-            .body("[0].originalPrice", equalTo(30000))
-            .body("[0].currentRecruitment", equalTo(3))
-            .body("[0].finalRecruitment", equalTo(3))
-            .body("[0].createdDateTime", equalTo("2022-08-03 16:00"))
-            .body("[0].closedDateTime", equalTo("2023-08-03 16:00"));
-    }
-
-    @Test
-    void 내가_등록한_쉐어를_조회한다() {
-        String accessToken = jwtProvider.createAccessToken(2370842997L);
-
-        given(documentationSpec)
-            .filter(document("share-search-mine-writer-get"))
-            .accept(APPLICATION_JSON_VALUE)
-            .header(AUTHORIZATION, accessToken)
-            .param("mineType", "writer")
-            .param("shareType", "delivery")
-            .param("isExpired", "false")
-
-            .when()
-            .get("/shares/mine")
-
-            .then()
-            .statusCode(HttpStatus.OK.value())
-            .body("", hasSize(1))
-            .body("[0].id", equalTo(1))
-            .body("[0].thumbnailUrl", containsString("https://"))
-            .body("[0].title", equalTo("강남역에서 떡볶이 먹을 사람 모집합니다."))
-            .body("[0].location", equalTo("강남역"))
-            .body("[0].price", equalTo(10000))
-            .body("[0].originalPrice", equalTo(30000))
-            .body("[0].currentRecruitment", equalTo(3))
-            .body("[0].finalRecruitment", equalTo(3))
-            .body("[0].createdDateTime", equalTo("2022-08-03 16:00"))
-            .body("[0].closedDateTime", equalTo("2023-08-03 16:00"));
-    }
-
-    @Test
-    void 내가_찜한_쉐어를_조회한다() {
-        String accessToken = jwtProvider.createAccessToken(2370842997L);
-
-        given(documentationSpec)
-            .filter(document("share-search-mine-wish-get"))
-            .accept(APPLICATION_JSON_VALUE)
-            .header(AUTHORIZATION, accessToken)
-            .param("mineType", "wish")
-            .param("shareType", "delivery")
-            .param("isExpired", "false")
-
-            .when()
-            .get("/shares/mine")
-
-            .then()
-            .statusCode(HttpStatus.OK.value())
-            .body("", hasSize(1))
-            .body("[0].id", equalTo(2))
-            .body("[0].thumbnailUrl", containsString("https://"))
-            .body("[0].title", equalTo("판교역에서 치킨 먹을 사람 모집합니다."))
-            .body("[0].location", equalTo("판교역"))
-            .body("[0].price", equalTo(7000))
-            .body("[0].originalPrice", equalTo(28000))
-            .body("[0].currentRecruitment", equalTo(3))
-            .body("[0].finalRecruitment", equalTo(4))
-            .body("[0].createdDateTime", equalTo("2022-07-03 16:00"))
-            .body("[0].closedDateTime", equalTo("2023-07-03 16:00"));
-    }
-
-    @Test
     void 특정_쉐어의_상세정보를_조회한다() {
         given(documentationSpec)
             .filter(document("share-detail-get"))
